MTP-1970: Be more conservative when deleting bank accounts
Check if a bank account record has no senders nor recipients before
deleting it.

This is similar to what it was done before but:
1. it keeps tracks of the `RecipientProfile`'s bank accounts
2. then deletes the `RecipientProfile`. So that it wouldn't be counted
   against the bank accont(s)'s recipients
3. and finally delete the bank accounts which have no senders or recipients

Previously we were trying to determine if a `BankAccount` was orphan
or not before deleting the `RecipientProfile`. This meant that a
`BankAccount` would have at least 1 recipient (the `RecipientProfile`
that would be deleted because it doesn't have any disbursements).

For this reason we were checking only if the BankAccount had any
senders (a `SenderProfile` has many BankAccounts through the
`BankAccountSenderDetails` model, although this was bank transfers
from the public which no longer happen).

However, this could **potentially** lead to the deletion of a
`BankAccount` linked to different `RecipientProfile`s (this is
a bit of an edge case and we're not actually sure if it's possible
in practice, but it is theoretically possible, hence the caution).

Because we delete the `RecipientProfile` in step 2 above now, we can
check that a `BankAccount` has no senders nor recipients now.
This should be more robust.

TODO: We'll need to do something similar when deleting a `SenderProfile`.
Because a `SenderProfile` has many `BankAccount` through `BankTransferSenderDetails`
and deleting the `SenderProfile` would delete the `BankTransferSenderDetails`
**but not** the `BankAccount` leading to orphan records in that table.

+++ mtp_api/apps/core/management/commands/delete_old_data.py
@@ -107,18 +107,20 @@ class Command(BaseCommand):
                 self.print_message(f'Deleting associated Events...', depth+2)
                 self.delete_events(recipient_profile.recipientprofileevent_set.all(), depth+3)
 
-                for recipient_details in recipient_profile.bank_transfer_details.all():
-                    bank_account = recipient_details.recipient_bank_account
-                    # NOTE: Checking only senders because we reached this point via the recipient
-                    #       BankTransferRecipientDetails will be cascade-deleted anyway when the RecipientProfile
-                    #       will be deleted, so no point in checking if bank_account.recipients is empty.
-                    if not bank_account.senders.exists():
-                        self.print_message(f'BankAccount {bank_account} has no senders, deleting...', depth+2)
-                        record_deleted = bank_account.delete()
-                        self.print_message(f'Records deleted: {record_deleted}.', depth+3)
+                bank_accounts_to_check = [
+                    recipient_details.recipient_bank_account
+
+                    for recipient_details in recipient_profile.bank_transfer_details.all()
+                ]
 
                 record_deleted = recipient_profile.delete()
                 self.print_message(f'Records deleted: {record_deleted}.', depth+2)
+
+                for bank_account in bank_accounts_to_check:
+                    if not bank_account.senders.exists() and not bank_account.recipients.exists():
+                        self.print_message(f'BankAccount {bank_account} has no senders nor recipients, deleting...', depth+2)
+                        record_deleted = bank_account.delete()
+                        self.print_message(f'Records deleted: {record_deleted}.', depth+3)
             else:
                 RecipientProfile.objects \
                     .filter(id=recipient_profile.id) \
